Move handling of keyworded arguments and wildcards to EBB
Why: At first the keyword and wildcard behaviors were implemented in
the BranchingPrompt. Although this worked, when transitionning to the
command line config solving, this code would have needed to be duplica
ted. It also puts all the handling of inputs inside the EBB which help
ensure that the BranchingPrompt only serves as an interface with no
logic.