Refactor node.map to use `rval, node = fct()`
Why:

Since the method is applied recursively and builds a similar tree
structure, it is much flexible if we allow the callable object `fct` to
redefine the relative of the current node (parent is move upwards,
children if moving downwards).

How:

The callable object `fct` is still passed the current node with the
relative (`fct(self, resulting_parent_node)` or
`fct(self, resulting_children_nodes)`), but it is now expected to return
the item **plus** the relative node which should be inserted in the
resulting tree.

This means that the function could decide to return a totally different
parent than the one passed to the function. This is necessary if the
structure of the resulting tree depends on the results.