Update `Experiment` based on new refers definition
Why:

The definition of `refers` is now:

 - `root_id`: Experiment id of the root experiment of the EVC tree
    This helps improve efficiency on db fetches since we can fetch a
    tree simply by specifying the root_id. It is also simple to keep
    coherent compared to a double-linked tree with (parent, children)
    links.
 - `parent_id`: Parent id of the current experiment
 - `adapter`: CompositeAdaptor definition to make trials from parent
    compatible with the current one and vice-versa.

How:

On creation of new experiments, `refers` is updated after the experiment
is added to database such that `refers[root_id]` can point to itself.

Note:

Because `refers` contain sub-dictionaries which are edited after
configuration (adapters instantiation), we need to use deep copies of
the configuration when configuring.