Add EVC node to `Experiment`
Why:

For modularity, an `Experiment` does not need to be part of EVC and can
be standalone.

How:

EVC trees can be built using _shallow_ definitions (only pointing to
name of experiments). Once the `item` of the shallow nodes are accessed,
and experiment view is created and is connected to the EVC node. Once
the connection is made, the experiment now have access to trials from
the entire tree. Those trials can be fetched using method
`Experiment.fetch_trials_tree`, a recursive version of
`Experiment.fetch_trials`. Note that both return a list of trials, not a
tree of trials.

Note:

Experiment status and statistics are computed only based on its own
trials, not those of the tree.