Fix race condition for exp registration
Why:

When a race condition occurs during experiment registration, it means
the process which losses the race must rebuild the configuration, this
time including the configuration of the experiment found in the
database. Current implementation was trying to rebuild the experiment
using the same configuration twice, meaning that after losing the race
condition it would try to register with a slightly different
configuration and generate a branching event.

How:

Catch the race condition (DuplicateKeyError) inside
`build_from(cmdargs)` rather than `build_from_config()` such that it can
rebuild the configuration from scratch.