Check if config is new but not exp
Why:

With the refactorisation of the experiment building process, there is
now 2 experiment being built, adding a new possible race condition.
There can be a new experiment registered before the local one is
instantiated and there can be a new experiment registered while the
local one is being configured. Only the latter case was handled.

How:

Datetime is only set at registration time. We can check at configuration
time if the experiment was instantiated from db, if yet, than the config
should also have been built from db, which means it should include the
datetime as well. If not, it means the config was built without
up-to-date information, most likely a race condition.