Add CommandLineChange and ScriptConfigChange
Why:

Just as for code change conflicts, modifications to the commandline call
or the user's script configuration file should be detected as conflicts.
As such, there should be adapters to connect the two experiments.

How:

Add CommandLineChange and ScriptConfigChange which follow the same logic
and structure as CodeChange. They will probably all diverge in the
future, when each of them gets more sophisticated than crude NOEFFECT,
UNSURE and BREAK types.