Make space and dimension builders non singletons
Why:

SpaceBuilder needs to build many different experiment spaces when using
the EVC. First, there is the conflicts, for which we need to create
spaces for old and new configuration. Second, there is the EVC trees,
for which we need to build all spaces when instantiating the
experiments. In the later case, at the time of generating command lines
in the consumer, the template of the space builder is the one from the
last experiment built from the EVC tree, now the actual experiment.

How:

Remove meta-class Singleton, and rebuild the template inside the
consumer to make sure its template is the one of the proper experiment.